Added cloud support for service catalogs
