+++ cfme/services/service_catalogs/ui.py
@@ -1,36 +1,28 @@
 from navmazing import NavigateToAttribute, NavigateToSibling
 
 from widgetastic.exceptions import NoSuchElementException
-from widgetastic.utils import deflatten_dict, Parameter, ParametrizedString, VersionPick
-from widgetastic.widget import ParametrizedView, Select, Text, View
-from widgetastic_patternfly import Button, Input, BootstrapSelect
+from widgetastic.widget import Select, Text, View
+from widgetastic_patternfly import Button
 
 from cfme.base import Server
 from cfme.base.login import BaseLoggedInPage
-from cfme.exceptions import ItemNotFound
 from cfme.services.requests import RequestsView
-from cfme.services.service_catalogs import ServiceCatalogs
+from cfme.services.service_catalogs import ServiceCatalogs, BaseOrderForm
 from cfme.utils.appliance import MiqImplementationContext
 from cfme.utils.appliance.implementations.ui import navigator, CFMENavigateStep, navigate_to, ViaUI
 from cfme.utils.blockers import BZ
-from cfme.utils.version import Version
 from cfme.utils.wait import wait_for
 from widgetastic_manageiq import Accordion, ManageIQTree
 
 
-class ServicesCatalogView(BaseLoggedInPage):
+class ServicesCatalogsView(BaseLoggedInPage):
+    title = Text("#explorer_title_text")
+
     @property
     def in_service_catalogs(self):
         return (
             self.logged_in_as_current_user and
             self.navigation.currently_selected == ['Services', 'Catalogs'])
-
-    @property
-    def is_displayed(self):
-        return (
-            self.in_service_catalogs and
-            self.configuration.is_displayed and
-            not self.catalogs.is_dimmed)
 
     @View.nested
     class service_catalogs(Accordion):  # noqa
@@ -38,96 +30,20 @@ class ServicesCatalogView(BaseLoggedInPage):
 
         tree = ManageIQTree()
 
-
-class OrderForm(ServicesCatalogView):
-    """Represents the order form of a service.
-    This form doesn't have a static set of elements apart from titles and buttons. In the most cases
-    the fields can be either regular inputs or dropdowns. Their locators depend on field names. In
-    order to find and fill required fields a parametrized view is used here. The keys of a fill
-    dictionary should match ids of the fields. For instance there is a field with such html
-    <input id="some_key"></input>, so a fill dictionary should look like that:
-    {"some_key": "some_value"}
-    """
-    title = Text('#explorer_title_text')
-    dialog_title = Text(
-        VersionPick({
-            Version.lowest(): ".//div[@id='main_div']//h3",
-            "5.9": ".//div[@id='main_div']//h2"
-        })
-    )
-
-    @ParametrizedView.nested
-    class fields(ParametrizedView):  # noqa
-        PARAMETERS = ("key",)
-        input = Input(id=Parameter("key"))
-        select = Select(id=Parameter("key"))
-        param_input = Input(id=ParametrizedString("param_{key}"))
-        dropdown = VersionPick({
-            Version.lowest(): BootstrapSelect(Parameter("key")),
-            "5.9": BootstrapSelect(locator=ParametrizedString(
-                ".//div[contains(@class, 'bootstrap-select') and "
-                "select[@id={key|quote}]]"))
-        })
-        param_dropdown = VersionPick({
-            Version.lowest(): BootstrapSelect(ParametrizedString("param_{key}")),
-            "5.9": BootstrapSelect(locator=ParametrizedString(
-                ".//div[contains(@class, 'bootstrap-select') and "
-                "select[@id='param_{key}']]"))
-        })
-
-        @property
-        def visible_widget(self):
-            if self.input.is_displayed:
-                return self.input
-            elif self.dropdown.is_displayed:
-                return self.dropdown
-            elif self.param_input.is_displayed:
-                return self.param_input
-            elif self.param_dropdown.is_displayed:
-                return self.param_dropdown
-            elif self.select.is_displayed:
-                return self.select
-            else:
-                raise ItemNotFound("Visible widget is not found")
-
-        def read(self):
-            return self.visible_widget.read()
-
-        def fill(self, value):
-            return self.visible_widget.fill(value)
-
-    def fill(self, fill_data):
-        values = deflatten_dict(fill_data)
-        was_change = False
-        self.before_fill(values)
-        for key, value in values.items():
-            widget = self.fields(key)
-            if value is None:
-                self.logger.debug('Skipping fill of %r because value was None', key)
-                continue
-            try:
-                if widget.fill(value):
-                    was_change = True
-            except NotImplementedError:
-                continue
-
-        self.after_fill(was_change)
-        return was_change
-
-
-class ServiceCatalogsView(ServicesCatalogView):
-    title = Text("#explorer_title_text")
-
     @property
     def is_displayed(self):
         return (
-            self.in_explorer and
+            self.in_service_catalogs and
             self.title.text == 'All Services' and
             self.service_catalogs.is_opened and
             self.service_catalogs.tree.currently_selected == ["All Services"])
 
 
-class ServiceCatalogsDefaultView(ServicesCatalogView):
+class OrderForm(ServicesCatalogsView, BaseOrderForm):
+    pass
+
+
+class ServiceCatalogsDefaultView(ServicesCatalogsView):
     title = Text("#explorer_title_text")
 
     @property
@@ -138,7 +54,7 @@ class ServiceCatalogsDefaultView(ServicesCatalogView):
             self.service_catalogs.is_opened)
 
 
-class DetailsServiceCatalogView(ServicesCatalogView):
+class DetailsServiceCatalogView(ServicesCatalogsView):
     title = Text("#explorer_title_text")
 
     order_button = Button("Order")
@@ -152,6 +68,7 @@ class DetailsServiceCatalogView(ServicesCatalogView):
 
 
 class OrderServiceCatalogView(OrderForm):
+    title = Text('#explorer_title_text')
     submit_button = Button('Submit')
 
     @property
@@ -180,7 +97,6 @@ def order(self):
     # TODO Remove once repaired
     if BZ(1513541, forced_streams=['5.9']).blocks:
         raise NotImplementedError("Service Order is broken - check BZ 1513541")
-    view.submit_button.click()
     # Additional check if submit is done, to cover case when last filled field value is not taken
     # Can appear every 2nd case when whole module is run
     if view.is_displayed and view.submit_button.is_displayed:
@@ -188,7 +104,6 @@ def order(self):
     view = self.create_view(RequestsView)
     view.flash.assert_no_error()
     view.flash.assert_message(msg, msg_type)
-    return self.appliance.collections.requests.instantiate(self.name, partial_check=True)
 
 
 @navigator.register(Server)
@@ -203,7 +118,7 @@ class ServiceCatalogsDefault(CFMENavigateStep):
 
 @navigator.register(ServiceCatalogs, 'All')
 class ServiceCatalogsAll(CFMENavigateStep):
-    VIEW = ServiceCatalogsView
+    VIEW = ServicesCatalogsView
 
     prerequisite = NavigateToAttribute('appliance.server', 'LoggedIn')
